Fix settings reload when a QStyleSheetStyle is in use
When an application sets a stylesheet, QApplication::style() no
longer returns the Qt6CTProxyStyle but rather a QStyleSheetStyle.
The existing logic for reloading style settings did not work in
this case.

Instead of creating a new Qt6CTProxyStyle and setting it as the
application style, we can take a different approach of reusing the
existing Qt6CTProxyStyle instances, and merely resetting their base
styles.  This works even when the Qt6CTProxyStyle is "hidden" under
a QStyleSheetStyle.

In order to track the Qt6CTProxyStyle instances, I have changed the
build to compile qt6ct.cpp into a shared "qt6ct-common" library.
This allows data in qt6ct.cpp to be shared between the platform
theme plugin and the style plugin.